modify Cross-origin resource sharing

+++ Backend/src/auth/auth.service.ts
@@ -28,23 +28,41 @@ export class AuthService {
       return { message: 'User with this email already exists' };
     }
 
-    return this.prisma.user.create({
+    const user = await this.prisma.user.create({
       data,
     });
+
+    return user;
   }
 
-  async update(id: number, data: { email?: string; name?: string; password?: string }): Promise<AuthDto> {
-    return this.prisma.user.update({
+  async update(id: number, data: { email?: string; name?: string; password?: string }): Promise<AuthDto|{message:string}> {
+    const { email, name, password } = data;
+    const existingUser = await this.prisma.user.findUnique({ where: { email } });
+
+    if (existingUser) {
+      return { message: 'User with this id is not found' };
+    }
+    const updateduser = await this.prisma.user.update({
       where: { id },
       data,
     });
+
+    return updateduser;
   }
 
-  async delete(id: number): Promise<AuthDto> {
-    return this.prisma.user.delete({
+  async delete(id: number): Promise<AuthDto|{message:string}> {
+    const existingUser = await this.prisma.user.findUnique({ where: {id} });
+
+    if (!existingUser) {
+      return { message: 'User with this id is not found' };
+    }
+    const deleteduser = await this.prisma.user.delete({
       where: { id },
     });
+    return deleteduser;
   }
+
+
   async signin(){
     return "it is working";
   }
